Move the helper to asylum.utils for reuse
