+++ project/members/handlers.py
@@ -2,6 +2,7 @@ import importlib
 import functools
 from django.conf import settings
 from django.db import transaction
+from asylum.util import get_handler_instance
 
 
 class BaseHandler(object):
@@ -34,21 +35,6 @@ class BaseApplicationHandler(BaseHandler):
 class BaseMemberHandler(BaseHandler):
     """Baseclass for callback handlers for Member processing"""
     pass
-
-
-def get_handler_instance(setting):
-    try:
-        setting_value = getattr(settings, setting)
-        #print("Got value %s for %s" % (setting_value, setting))
-    except AttributeError:
-        #print("Got AttributeError for %s" % (setting))
-        return None
-    if not setting_value:
-        return None
-    module_name, class_name = setting_value.rsplit(".", 1)
-    HandlerClass = getattr(importlib.import_module(module_name), class_name)
-    instance = HandlerClass()
-    return instance
 
 
 def call_saves(setting):
